Rename: 컴포넌트 파일 이동, 스토리 파일 이동
